# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WebChat Public API
  version: 0.4.0
  description: >
    Публичный API веб-чата. Единственный источник истины публичного API;
    TypeScript-типы фронтенда генерируются из этого файла автоматически.

    Фича 002 (auth): endpoints №1–10 — регистрация с email-подтверждением,
    вход по паролю, refresh-ротация, logout, восстановление пароля, /users/me.
    Публичны ровно endpoints №1–6, №8–9 (8 путей); №7 logout и №10 /users/me
    требуют Bearer. Все ответы об ошибках — RFC 9457 application/problem+json
    (схема Problem). Ссылки в письмах указывают на маршруты SPA, а не на API.

    Фича 003 (sso): endpoints SSO §1–7 — вход через внешних
    identity-провайдеров (OIDC/OAuth2): перечень провайдеров,
    authorize/callback/token, привязка и отвязка идентичностей. Дополнительно
    публичны 4 маршрута (providers, authorize, callback, token);
    link/authorize, GET /users/me/identities и DELETE идентичности
    требуют Bearer. Endpoint'ы 002 не изменились (аддитивно, minor-бамп).

    Фича 004 (messaging): endpoints №11–24 — личные текстовые диалоги
    (ensure/чтение/удаление, история с курсорной пагинацией, отправка с
    клиентской дедупликацией, read-отметки), реалтайм-канал SSE
    №18 /users/me/events, контакты, точный поиск пользователей и
    блокировки (№19–24). Все новые пути требуют Bearer: публичный
    перечень 002–003 не расширен. Фиксированные константы: максимум текста
...
    только additive (0.3.0 → 0.4.0, без BREAKING.md).
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/chats/ensure:
    post:
      operationId: ensureChat
      tags: [chats]
      summary: Открыть личный диалог с пользователем (№11, Bearer)
      description: >
        Возвращает существующий диалог пары пользователей либо создаёт новый
        (201 — создан; 200 — уже существует). Повторное открытие существующего
        чата снимает hidden у вызывающего: чат возвращается в список; история,
        удалённая через №14, остаётся скрытой водяным знаком deleted_up_to_seq.
        Пара пользователей имеет ровно один диалог (idempotent ensure, FR-001).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnsureChatRequest'
      responses:
        '200':
          description: Диалог уже существует — возвращён существующий ChatView (hidden снят у вызывающего)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatView'
        '201':
          description: Диалог создан — ChatView
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatView'
        '400':
          description: 'Невалидное тело (errors: {peerUserId: [invalid_uuid]})'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          description: Не аутентифицирован (нет токена / истёк / отозван / недействителен — единообразно)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: 'Собеседник не найден (errors: {peer: [peer_not_found]})'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: 'Диалог с самим собой запрещён (errors: {peerUserId: [self_forbidden]})'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/chats:
    get:
      operationId: listChats
      tags: [chats]
      summary: Список диалогов пользователя (№12, Bearer)
      description: >
        ChatListItem одним запросом. Серверная сортировка: последнее видимое
        сообщение (входящее или исходящее) по убыванию createdAt; чаты без
        видимых сообщений — ниже; исключены только удалённые без новых
        сообщений (hidden ∧ last_seq ≤ deleted_up_to_seq). Новый входящий
        снимает hidden и возвращает чат в список (FR-021). Превью
        lastMessage.text — полный текст (обрезка ≤64 симв. — клиентский
        рендер).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Диалоги пользователя, отсортированные по последнему видимому сообщению
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [chats]
                properties:
                  chats:
                    type: array
                    description: По убыванию createdAt последнего видимого сообщения; пустой список валиден
                    items:
                      $ref: '#/components/schemas/ChatListItem'
        '401':
          description: Не аутентифицирован (нет токена / истёк / отозван / недействителен — единообразно)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/chats/{chatId}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Диалог по идентификатору (№13, Bearer)
      description: >
        ChatView пары с водяными знаками прочтения. Авторизация на уровне
        ресурса — membership диалога (FR-002): не-участник получает 403
        not_participant на существующем чате; несуществующий — 404.
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          description: Идентификатор диалога
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Диалог
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatView'
        '401':
          description: Не аутентифицирован (нет токена / истёк / отозван / недействителен — единообразно)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: 'Вызывающий — не участник диалога (errors: {chat: [not_participant]})'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: 'Диалог не найден (errors: {chat: [chat_not_found]})'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      operationId: deleteChat
      tags: [chats]
      summary: Удаление диалога у себя (№14, Bearer)
      description: >
        Per-user удаление (FR-021): история скрывается водяным знаком
        deleted_up_to_seq = last_seq и hidden = true только у удалившего;
        второй участник не затронут. Новый входящий снимает hidden и
        возвращает чат в список — без старой истории (она остаётся за водяным
        знаком). Повторный вызов — 204 (идемпотентно).
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          description: Идентификатор диалога
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: История скрыта у вызывающего (идемпотентно)
        '401':
          description: Не аутентифицирован (нет токена / истёк / отозван / недействителен — единообразно)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: 'Вызывающий — не участник диалога (errors: {chat: [not_participant]})'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: 'Диалог не найден (errors: {chat: [chat_not_found]})'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /api/v1/users/{userId}/block:
    put:
      operationId: blockUser
      tags: [contacts]
      summary: Заблокировать пользователя (№23, Bearer)
      description: >
        Односторонняя блокировка, управляется блокирующим (FR-020).
        Идемпотентно: повторная блокировка — 204 без изменений (edge).
        Пока блок действует: отправка в диалоге пары запрещена в обе
        стороны без создания записей (№16 → 403 chat_blocked_by_you у
        блокирующего, 403 you_are_blocked у заблокированного — единственное
        явное уведомление заблокированного); прочтения блокирующего не
        двигают водяной знак и не публикуются, прочтения заблокированного
        двигают водяной знак локально без публикации (№17); бейдж
        непрочитанных у блокирующего «замирает» (№12). Контакты не
        затрагиваются. Единственная проекция блокировки в API —
        blockedByMe; поле «кто заблокировал меня» отсутствует. Разблокировка
        — №24.
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: Блокируемый пользователь
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Блокировка установлена (идемпотентно — повторная блокировка без изменений)
        '401':
          description: Не аутентифицирован (нет токена / истёк / отозван / недействителен — единообразно)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: 'Пользователь не найден (errors: {user: [user_not_found]})'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: 'Блокировка самого себя запрещена (errors: {userId: [self_forbidden]})'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      operationId: unblockUser
      tags: [contacts]
      summary: Разблокировать пользователя (№24, Bearer)
      description: >
        Снимает блокировку вызывающего с пользователя. Идемпотентно:
        разблокировка не заблокированного — тоже 204. Возвращает обычное
        поведение диалога — отправку, прочтения, бейдж и события прочтения
        (FR-010) — без потери истории (FR-020).
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: Разблокируемый пользователь
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Блокировка снята (идемпотентно); обычное поведение диалога возобновляется
        '401':
          description: Не аутентифицирован (нет токена / истёк / отозван / недействителен — единообразно)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Authorization: Bearer <accessToken>'
  schemas:
...
    PublicUser:
      type: object
      additionalProperties: false
      required: [id, username, email, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[a-zA-Z0-9]([a-zA-Z0-9_.-]{1,30}[a-zA-Z0-9])$'
        email:
          type: string
          format: email
          maxLength: 254
        status:
          type: string
          enum: [pending_email_confirmation, awaiting_password, active]
        createdAt:
          type: string
          format: date-time
    Problem:
      type: object
      description: 'RFC 9457: единая модель ошибок; расширения допускаются'
      required: [title, status]
      properties:
        type:
          type: string
          description: URI-reference на тип ошибки (опционально)
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: object
          description: Коды ошибок валидации по полям (в 400-ответах)
          additionalProperties:
            type: array
            items:
              type: string
              description: Код ошибки, например invalid_format
...
    EnsureChatRequest:
      type: object
      additionalProperties: false
      required: [peerUserId]
      properties:
        peerUserId:
          type: string
          format: uuid
          description: Собеседник; диалог с самим собой → 422 self_forbidden
    ChatView:
      type: object
      additionalProperties: false
      required: [chatId, peer, blockedByMe, peerReadUpToSeq, myReadUpToSeq]
      properties:
        chatId:
          type: string
          format: uuid
          description: Идентификатор диалога пары пользователей
        peer:
          $ref: '#/components/schemas/PublicUser'
          description: Собеседник (вторая сторона диалога)
        blockedByMe:
          type: boolean
          description: 'Вызывающий блокирует собеседника (FR-020). Единственная проекция блокировок в API: поле «кто заблокировал меня» отсутствует — заблокированный узнаёт о блокировке только из 403 you_are_blocked при отправке'
        peerReadUpToSeq:
          type: integer
          format: int64
          minimum: 0
          description: Водяной знак прочтения собеседника (0 — ничего не прочитано); для двойной галочки отправителя
        myReadUpToSeq:
          type: integer
          format: int64
          minimum: 0
          description: Водяной знак прочтения вызывающего (0 — ничего не прочитано)
    Message:
      type: object
      additionalProperties: false
      required: [id, chatId, senderId, text, seq, createdAt]
      properties:
        id:
          type: string
          format: uuid
          description: '= clientMessageId из №16: клиентский UUID, exactly-once дедупликация (FR-004)'
        chatId:
          type: string
          format: uuid
        senderId:
          type: string
          format: uuid
          description: Отправитель (участник диалога)
        text:
          type: string
          maxLength: 4096
          description: Триммированный по краям текст (FR-003); внутренние пробелы сохраняются; ≤4096 символов после trim
        seq:
          type: integer
          format: int64
          minimum: 1
          description: Монотонный порядок внутри диалога (уникален в рамках чата); основа курсорной пагинации №15
        createdAt:
          type: string
          format: date-time
    ChatListItem:
      type: object
      additionalProperties: false
      required: [chatId, peer, lastMessage, unreadCount, blockedByMe]
      properties:
        chatId:
          type: string
          format: uuid
        peer:
          $ref: '#/components/schemas/PublicUser'
          description: Собеседник
        lastMessage:
          description: Последнее видимое вызывающему сообщение (входящее или исходящее); null — видимых сообщений нет (пустой или полностью удалённый у себя диалог)
          anyOf:
            - $ref: '#/components/schemas/Message'
            - type: 'null'
        unreadCount:
          type: integer
          format: int64
          minimum: 0
          description: Входящие сообщения с seq > myReadUpToSeq (видимые вызывающему); при блокировке у блокирующего «замирает» (FR-020)
        blockedByMe:
          type: boolean
          description: Вызывающий блокирует собеседника — метка «заблокирован» в UI (FR-020)
...